convert stream logic to posts logic
